server updated

+++ src/routes/memories.ts
@@ -3,8 +3,15 @@ import { prisma } from "../lib/prisma";
 import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
+
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: { createdAt: "asc" },
     });
     return memories.map((memory) => {
@@ -12,11 +19,12 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id: memory.id,
         coverUrl: memory.coverUrl,
         excerpt: memory.content.substring(0, 115).concat("..."),
+        createdAt: memory.createdAt,
       };
     });
   });
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, response) => {
     const paramSchema = z.object({
       id: z.string().uuid(),
     });
@@ -25,6 +33,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const memory = await prisma.memory.findUniqueOrThrow({
       where: { id },
     });
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return response.status(401).send();
+    }
 
     return memory;
   });
@@ -37,19 +49,20 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
+
     const memory = await prisma.memory.create({
       data: {
         content,
         coverUrl,
         isPublic,
-        userId: "a954061e-fd76-4e30-8657-2c968d5e8135",
+        userId: request.user.sub,
       },
     });
 
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, response) => {
     const paramSchema = z.object({
       id: z.string().uuid(),
     });
@@ -63,7 +76,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-    await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: { id },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return response.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: { id },
       data: {
         content,
@@ -71,13 +92,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
         isPublic,
       },
     });
+    return memory;
   });
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, response) => {
     const paramSchema = z.object({
       id: z.string().uuid(),
     });
     const { id } = paramSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: { id },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return response.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: { id },
